--- a/day_2/solution.go
+++ b/day_2/solution.go
@@ -2,118 +2,327 @@ package day2
 
 import (
 	"github.com/pivovarit/aoc/util"
-	"log"
-	"math"
-	"strconv"
-	"strings"
-)
-
-const (
-	BLUE  = "blue"
-	GREEN = "green"
-	RED   = "red"
+	"runtime"
+	"sync"
 )
 
 func run() {
 	input := util.ReadInput()
 
-	util.Timed("cubeConundrumPart1", func() int {
-		return cubeConundrumPart1(input)
+	util.Timed("cubeConundrumPart1Sequential", func() int {
+		return cubeConundrumPart1Sequential(input)
 	})
-	util.Timed("cubeConundrumPart1", func() int {
-		return cubeConundrumPart2(input)
+	util.Timed("cubeConundrumPart1Parallel", func() int {
+		return cubeConundrumPart1Parallel(input)
+	})
+	util.Timed("cubeConundrumPart2Sequential", func() int {
+		return cubeConundrumPart2Sequential(input)
+	})
+
+	util.Timed("cubeConundrumPart2Parallel", func() int {
+		return cubeConundrumPart2Parallel(input)
 	})
 }
 
-func cubeConundrumPart1(input []string) int {
-	colourMap := map[string]int{
-		RED:   12,
-		GREEN: 13,
-		BLUE:  14,
-	}
+const (
+	redCubesCount   = 12
+	greenCubesCount = 13
+	blueCubesCount  = 14
+)
 
-	idSum := 0
-	for _, str := range input {
-		id, rounds := parseEntry(str)
-		invalid := false
-		for _, round := range rounds {
-			for _, choice := range round.choices {
-				if choice.count > colourMap[choice.colour] {
-					invalid = true
-					break
+type State int
+
+const (
+	processingGameId State = iota
+	processingCubeCount
+	processingCubeColour
+)
+
+const (
+	emptySeparator      = ' '
+	gameIdSeparator     = ':'
+	cubeChoiceSeparator = ','
+	gameSeparator       = ';'
+)
+
+func cubeConundrumPart1Sequential(input []string) (sum int) {
+LineLoop:
+	for _, line := range input {
+		state := processingGameId
+		gameId := 0
+		cubeCount := 0
+		cubeColourProcessed := false
+		colourPrefix := true
+		for _, char := range line[5:] {
+			switch state {
+			case processingGameId:
+				if char != gameIdSeparator {
+					gameId = gameId*10 + int(char-'0')
+				} else {
+					state = processingCubeCount
+				}
+			case processingCubeCount:
+				switch {
+				case char == emptySeparator && colourPrefix:
+					colourPrefix = false
+
+				case char != emptySeparator:
+					cubeCount = cubeCount*10 + int(char-'0')
+
+				case char == emptySeparator && !colourPrefix:
+					state = processingCubeColour
+				}
+			case processingCubeColour:
+				if char != cubeChoiceSeparator && char != gameSeparator {
+					if cubeColourProcessed {
+						continue
+					}
+					switch char {
+					case 'r':
+						if cubeCount > redCubesCount {
+							continue LineLoop
+						}
+					case 'b':
+						if cubeCount > blueCubesCount {
+							continue LineLoop
+						}
+					case 'g':
+						if cubeCount > greenCubesCount {
+							continue LineLoop
+						}
+					}
+					cubeColourProcessed = true
+				} else {
+					cubeColourProcessed = false
+					cubeCount = 0
+					state = processingCubeCount
+					colourPrefix = true
 				}
 			}
 		}
-		if !invalid {
-			idSum += id
-		}
+		sum += gameId
 	}
-	return idSum
+	return sum
 }
 
-type GameRound struct {
-	choices []GameEntry
-}
+func cubeConundrumPart1Parallel(input []string) (sum int) {
+	var wg sync.WaitGroup
+	parallelism := runtime.NumCPU()
+	inputLen := len(input)
+	parallelism = min(inputLen, parallelism)
 
-type GameEntry struct {
-	count  int
-	colour string
-}
+	var idSumPerCPU = make([]int, parallelism)
 
-func Entry(count int, colour string) GameEntry {
-	return GameEntry{
-		count:  count,
-		colour: colour,
-	}
-}
+	for cpuId := 0; cpuId < parallelism; cpuId++ {
+		wg.Add(1)
+		go func(cpu int) {
+			defer wg.Done()
+		LineLoop:
+			for i := cpu; i < inputLen; i = i + parallelism {
+				line := input[i]
+				state := processingGameId
+				gameId := 0
+				cubeCount := 0
+				cubeColourProcessed := false
+				colourPrefix := true
+				for _, char := range line[5:] {
+					switch state {
+					case processingGameId:
+						if char != gameIdSeparator {
+							gameId = gameId*10 + int(char-'0')
+						} else {
+							state = processingCubeCount
+						}
+					case processingCubeCount:
+						if char == emptySeparator && colourPrefix {
+							colourPrefix = false
+							break
+						}
 
-func Round(entries ...GameEntry) GameRound {
-	return GameRound{choices: entries}
-}
+						if char != emptySeparator {
+							cubeCount = cubeCount*10 + int(char-'0')
+							break
+						}
 
-func parseEntry(input string) (int, []GameRound) {
-	split := strings.Split(input, ":")
-	id, err := strconv.Atoi(split[0][5:])
-	if err != nil {
-		log.Panicf("illegal input format")
-	}
-
-	var rounds []GameRound
-
-	rawRounds := strings.Split(strings.TrimSpace(split[1]), ";")
-
-	for _, round := range rawRounds {
-		var entries []GameEntry
-		rawEntries := strings.Split(round, ",")
-		for _, entry := range rawEntries {
-			parsedEntry := strings.Split(strings.TrimSpace(entry), " ")
-			count, err := strconv.Atoi(strings.TrimSpace(parsedEntry[0]))
-			if err != nil {
-				panic("illegal game round format")
+						if char == emptySeparator && !colourPrefix {
+							state = processingCubeColour
+						}
+					case processingCubeColour:
+						if char != cubeChoiceSeparator && char != gameSeparator {
+							if cubeColourProcessed {
+								continue
+							}
+							switch char {
+							case 'r':
+								if cubeCount > redCubesCount {
+									continue LineLoop
+								}
+							case 'b':
+								if cubeCount > blueCubesCount {
+									continue LineLoop
+								}
+							case 'g':
+								if cubeCount > greenCubesCount {
+									continue LineLoop
+								}
+							}
+							cubeColourProcessed = true
+						} else {
+							cubeColourProcessed = false
+							cubeCount = 0
+							state = processingCubeCount
+							colourPrefix = true
+						}
+					}
+				}
+				idSumPerCPU[cpu] += gameId
 			}
-
-			colour := strings.TrimSpace(parsedEntry[1])
-
-			entries = append(entries, Entry(count, colour))
-		}
-
-		rounds = append(rounds, Round(entries...))
+		}(cpuId)
 	}
 
-	return id, rounds
+	wg.Wait()
+
+	for _, sumPerCpu := range idSumPerCPU {
+		sum += sumPerCpu
+	}
+
+	return sum
 }
 
-func cubeConundrumPart2(input []string) int {
-	sum := 0
-	for _, str := range input {
-		maxCubes := map[string]int{}
-		_, rounds := parseEntry(str)
-		for _, round := range rounds {
-			for _, entry := range round.choices {
-				maxCubes[entry.colour] = int(math.Max(float64(maxCubes[entry.colour]), float64(entry.count)))
+func cubeConundrumPart2Sequential(input []string) (sum int) {
+	for _, line := range input {
+		state := processingGameId
+		gameId := 0
+		cubeCount := 0
+		cubeColourProcessed := false
+		colourPrefix := true
+		maxGreen := -1
+		maxBlue := -1
+		maxRed := -1
+		for _, char := range line[5:] {
+			switch state {
+			case processingGameId:
+				if char != gameIdSeparator {
+					gameId = gameId*10 + int(char-'0')
+				} else {
+					state = processingCubeCount
+				}
+			case processingCubeCount:
+				if char == emptySeparator && colourPrefix {
+					colourPrefix = false
+					break
+				}
+
+				if char != emptySeparator {
+					cubeCount = cubeCount*10 + int(char-'0')
+					break
+				}
+
+				if char == emptySeparator && !colourPrefix {
+					state = processingCubeColour
+				}
+			case processingCubeColour:
+				if char != cubeChoiceSeparator && char != gameSeparator {
+					if cubeColourProcessed {
+						continue
+					}
+					switch char {
+					case 'r':
+						maxRed = max(maxRed, cubeCount)
+					case 'b':
+						maxBlue = max(maxBlue, cubeCount)
+					case 'g':
+						maxGreen = max(maxGreen, cubeCount)
+					}
+					cubeColourProcessed = true
+				} else {
+					cubeColourProcessed = false
+					cubeCount = 0
+					state = processingCubeCount
+					colourPrefix = true
+				}
 			}
 		}
-		sum += maxCubes[RED] * maxCubes[GREEN] * maxCubes[BLUE]
+		sum += maxBlue * maxGreen * maxRed
 	}
+	return sum
+}
+
+func cubeConundrumPart2Parallel(input []string) (sum int) {
+	var wg sync.WaitGroup
+	parallelism := runtime.NumCPU()
+	inputLen := len(input)
+	parallelism = min(inputLen, parallelism)
+
+	var powerPerCPU = make([]int, parallelism)
+
+	for cpuId := 0; cpuId < parallelism; cpuId++ {
+		wg.Add(1)
+		go func(cpu int) {
+			defer wg.Done()
+			for i := cpu; i < inputLen; i = i + parallelism {
+				line := input[i]
+				state := processingGameId
+				gameId := 0
+				cubeCount := 0
+				cubeColourProcessed := false
+				colourPrefix := true
+				maxGreen := -1
+				maxBlue := -1
+				maxRed := -1
+				for _, char := range line[5:] {
+					switch state {
+					case processingGameId:
+						if char != gameIdSeparator {
+							gameId = gameId*10 + int(char-'0')
+						} else {
+							state = processingCubeCount
+						}
+					case processingCubeCount:
+						if char == emptySeparator && colourPrefix {
+							colourPrefix = false
+							break
+						}
+
+						if char != emptySeparator {
+							cubeCount = cubeCount*10 + int(char-'0')
+							break
+						}
+
+						if char == emptySeparator && !colourPrefix {
+							state = processingCubeColour
+						}
+					case processingCubeColour:
+						if char != cubeChoiceSeparator && char != gameSeparator {
+							if cubeColourProcessed {
+								continue
+							}
+							switch char {
+							case 'r':
+								maxRed = max(maxRed, cubeCount)
+							case 'b':
+								maxBlue = max(maxBlue, cubeCount)
+							case 'g':
+								maxGreen = max(maxGreen, cubeCount)
+							}
+							cubeColourProcessed = true
+						} else {
+							cubeColourProcessed = false
+							cubeCount = 0
+							state = processingCubeCount
+							colourPrefix = true
+						}
+					}
+				}
+				powerPerCPU[cpu] += maxBlue * maxGreen * maxRed
+			}
+		}(cpuId)
+	}
+	wg.Wait()
+
+	for _, sumPerCpu := range powerPerCPU {
+		sum += sumPerCpu
+	}
+
 	return sum
 }