Before:
```
CubeConundrumPart1-10              7238            166293 ns/op          154312 B/op       3380 allocs/op
CubeConundrumPart2-10              6255            187504 ns/op          154312 B/op       3380 allocs/op
```
After:
```
CubeConundrumPart1Sequential-10   80180             14467 ns/op               0 B/op          0 allocs/op
CubeConundrumPart1Parallel-10    105219             10908 ns/op            1144 B/op         23 allocs/op
CubeConundrumPart2Sequential-10   57236             20915 ns/op               0 B/op          0 allocs/op
CubeConundrumPart2Parallel-10     75871             15214 ns/op            1144 B/op         23 allocs/op
```